Fix test and break out 8.13-specific test coverage to separate cross-version test class
