+++ platforms/core-configuration/configuration-cache/src/crossVersionTest/groovy/org/gradle/internal/cc/impl/ConfigurationCacheProblemsTapi813PlusCrossVersionTest.groovy
@@ -29,9 +29,9 @@ import org.gradle.tooling.events.problems.Severity
 import org.gradle.tooling.events.problems.SingleProblemEvent
 import org.gradle.tooling.events.problems.internal.GeneralData
 
-@ToolingApiVersion(">=8.12 <8.13")
+@ToolingApiVersion(">=8.13")
 @TargetGradleVersion(">=8.10")
-class ConfigurationCacheProblemsTapi812PlusCrossVersionTest extends ToolingApiSpecification {
+class ConfigurationCacheProblemsTapi813PlusCrossVersionTest extends ToolingApiSpecification {
 
     class ProblemProgressListener implements ProgressListener {
 
@@ -82,8 +82,8 @@ class ConfigurationCacheProblemsTapi812PlusCrossVersionTest extends ToolingApiSp
             definition.id.group.displayName == "configuration cache validation"
             definition.id.group.name == "configuration-cache"
             definition.severity == Severity.ERROR
-            (locations[0] as LineInFileLocation).path == "build file 'build.gradle'" // FIXME: the path should not contain a prefix nor extra quotes
-            (locations[1] as LineInFileLocation).path == "build file '$buildFile.path'"
+            (originLocations[0] as LineInFileLocation).path == "build file 'build.gradle'" // FIXME: the path should not contain a prefix nor extra quotes
+            (originLocations[1] as LineInFileLocation).path == "build file '$buildFile.path'"
             additionalData instanceof GeneralData
             additionalData.asMap.isEmpty()
         }
